Separate settings file
